Validate: improve performance (#164)

* significantly reduce serde_json serialization by only retrieving validation output for invalid rows, and directly retrieving needed fields instead of traversing JSON tree
* change validation error output format from jsonl to TSV. This avoids json manipulation just to insert row number, with added bonus of making output dramatically easier to read
* performance improved by ~2.6X (from ~36s to ~14s)

```
(base) ➜  qsv git:(validate_performance) ✗ time ./target/release/qsvlite validate tmp/NYC_311_SR_2010-2020-sample-1M.csv tmp/NYC-short.csv.schema.json
[00:00:13] [==================== 100% validated 1,000,000 records.] (126,103/sec)
6,494 out of 1,000,000 records invalid.
./target/release/qsvlite validate tmp/NYC_311_SR_2010-2020-sample-1M.csv   25.45s user 0.84s system 189% cpu 13.841 total

```

*validation error file for adur public toilets*
```
row_number      field   error
1       ExtractDate     null is not of type "string"
1       OrganisationLabel       null is not of type "string"
3       CoordinateReferenceSystem       "OSGB3" does not match "(WGS84|OSGB36)"
3       Category        "Mens" does not match "(Female|Male|Female and Male|Unisex|Male urinal|Children only|None)"
```

flamegraph looks much cleaner

![flamegraph](https://user-images.githubusercontent.com/9694055/155090772-b3099324-544a-462d-b8a4-388d1added7a.svg)