@jqnatividad 2022-02-22
@mhuang74 reducing serde serialization was spot on and changing the report format was inspired - increasing readability AND increasing performance at the same time!